requirements.txt
absl-py==0.7.1
APScheduler==3.5.1
asn1crypto==0.24.0
astor==0.7.1
attrs==19.1.0
Automat==0.7.0
blinker==1.4
boto3==1.9.114
botocore==1.12.114
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
colorclass==2.2.0
coloredlogs==10.0
colorhash==1.0.2
ConfigArgParse==0.13.0
constantly==15.1.0
cryptography==2.6.1
cycler==0.10.0
cymem==2.0.2
cytoolz==0.9.0.1
decorator==4.3.2
dill==0.2.9
docopt==0.6.2
docutils==0.14
en-core-web-md==2.0.0
en-core-web-sm==2.0.0
fakeredis==0.10.3
fbmessenger==5.3.2
Flask==1.0.2
Flask-Cors==3.0.7
Flask-JWT-Simple==0.0.3
future==0.17.1
gast==0.2.2
gevent==1.4.0
greenlet==0.4.15
grpcio==1.19.0
h5py==2.7.1
humanfriendly==4.18
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.4
jsonpickle==1.0
jsonschema==2.6.0
kafka-python==1.4.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
klein==17.10.0
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==2.2.4
mattermostwrapper==2.1
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.2
networkx==2.2
numpy==1.16.0
packaging==18.0
pandas==0.24.2
pika==0.12.0
plac==0.9.6
preshed==2.0.1
prompt-toolkit==2.0.9
protobuf==3.7.0
pycparser==2.19
pydot==1.4.1
PyHamcrest==1.9.0
PyJWT==1.7.1
pykwalify==1.7.0
pymongo==3.7.2
pyparsing==2.3.1
PySocks==1.6.8
python-crfsuite==0.9.6
python-dateutil==2.7.5
python-engineio==3.4.4
python-socketio==3.1.1
python-telegram-bot==11.1.0
pytz==2018.9
PyYAML==5.1
questionary==1.0.2
rasa-core==0.13.3
rasa-core-sdk==0.12.2
rasa-nlu==0.14.4
redis==2.10.6
regex==2018.1.10
requests==2.21.0
requests-toolbelt==0.9.1
rocketchat-API==0.6.25
ruamel.yaml==0.15.85
s3transfer==0.2.0
scikit-learn==0.20.2
scipy==1.2.0
simplejson==3.16.0
six==1.11.0
sklearn-crfsuite==0.3.6
slackclient==1.3.0
spacy==2.0.18
tabulate==0.8.3
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
terminaltables==3.1.0
thinc==6.12.1
toolz==0.9.0
tqdm==4.29.1
twilio==6.23.0
Twisted==18.9.0
typing==3.6.6
tzlocal==1.5.1
ujson==1.35
urllib3==1.24.1
wcwidth==0.1.7
webexteamssdk==1.1.1
websocket-client==0.54.0
Werkzeug==0.14.1
wrapt==1.10.11
zope.interface==4.6.0
